[REFACTOR] Rename "linux" module to "unix"
If we're going to conditionally compile something with `cfg(unix)`,
it's only confusing to call the compiled module `linux`.
